pyproject.toml
[project]
name = "rime-phah-taibun"
version = "0.1.0"
description = "拍台文 — 基於 Rime 的台語輸入法方案，資料前處理工具"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14",
    "pytest>=8.0",
    "pytest-cov>=6.0",
]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific
]

[tool.ruff.lint.isort]
known-first-party = ["scripts"]

[tool.pytest.ini_options]
testpaths = ["tests"]
